Import the XPCS and averaging libraries.
XPCS is designed for photon correlation experiments. Currently it calculates the g2 correlation function for many different normalizations.

averaging is a support library designed for spatial averaging of an image.  For example, if one wanted to smooth out a speckle image to create an incoherent background, these functions are used.  This is used quite a bit for g2 functions where q-averaging is used in many of the normalization schemes.