gpu_domain_generator:
Initial commit of code to do labyrinth domain generation on the gpu. I don't know how useful this will be for anybody else but it might be helpful to see examples of calling the gpu with python/pyopencl. Later I will add example code showing how to use the class because the simulation must be run in specific ways.

kernels: opencl/c99 kernels. Right now these are just docstrings which get read and compiled by the opencl runtime. A better organizational method is desired (separate files for each kernel in another directory?).