propagate: 1. Add option to band limit the propagated field to the upper limit of the propagator's reliability 2. Change behavior of acutance to allow a set of frames to be submitted at the same time. Why it wasnt this way before I have no idea.
averaging: add a new type of smoothing that I found somewhere. I use it for my financial stuff, it gives pretty results.

gpu_correlations: many, many changes to allow different array sizes for the following three quantities: i. the object being probed ii. the probe/illumination function iii. the simulated speckle pattern. gpu_correlations is now out-of-sync with the cpu code base, but since I'm the only one who uses the functions this is low priority

shape: added circular fermi-dirac function to use as a low-pass filter in propagate

gpu_domains: modifications to the envelope builder so that the file names can be more descriptive (there is a now a class variable which builds up a string to describe the envelope function)

gpu_phasing: add prtf function to implement the phase-retrieval transfer function calculation. basically, this function calculates the reproducibility of the phase as a function of q. Resolution of the reconstruction is determined from this function.