CL buffers: various
gpu: add method to allow CPU to be instantiated as a CL device if the Intel OCL runtime is installed. Fix problem where creating context would hang the tesla; the fix is to wait 0.5 seconds between querying devices and context creation. This may be a leading indicator of a hardware failure. as its appearance was sudden and unresolved by reinstallation of the nVidia CUDA Toolkit 5.0 and updated driver.

gpu_correlations: include simulation of beamblock with variable size for the stick (symm micr sims). kernels to implement += and *= operations. on-gpu despiking (was this there before?). include memory buffers so that the total corr and total cosine decomp can be returned (useful for average spectra).

gpu_propagate: big changes in how subarray is handled to allow for rectangular subregions (subarray refers to the region saved for output). This required rewriting copy_to_save_buffer.cl.

phasing: improve rftf. make align_global_phase more robust.

symmetries: make api more compatible with the gpu code