new dust removal function and corresponding plan function which uses griddata to do an interpolation from irregularly spaced data. it is unclear if this new function performs better or faster, but the code should be easier to understand than the old version
gpu_phasing: big cleanup and refactoring of load_data with an eye towards fully turning on the partial coherence correction in clark APL 99 2011. To improve readability of the code, many of the fragmentary kernels were reimplemented as external files. The basic_*.cl files are rudimentary mathematical operations such as division of complex numbers (f indicates float, f2 float2 or complex).