in conditioning, change behavior of subtract_dark to allow use of a mask for matching dark counts.
in fit, add a dictionary to the output so that the calculated attributes may be easily queried.

in gpu, change behavior of the start() method so that it can accept an existing context and queue

in io, change pickle to cpickle when available. add complexhsv which duplicates complex_hsv, but in some instances it is better to not have _ in the component name.

in domain_generator and symmetry_microscope, update __init__() to conform to new gpu.start() standards.

in symmetry_micrscope, include a new method for calculating concentration

in kernels: various new kernels and a bug fix in sm_correl_norm (prefer hypot over native_sqrt)